fix: Make engine an actual, usable object
In the original Udacity version of engine.js, it says it is
supposed to return the Engine object as a game engine
"controller" with functions that can be accessed, but what is
actually returned is an "undefined" value from the closure.
I added a return value that returns all the functions from Engine
so they can be accessed from the global scope.